requirements.txt
aniso8601==8.0.0
asn1crypto==1.3.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
docutils==0.15.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-RESTful==0.3.8
future==0.18.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.4
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
psutil==5.6.7
pycosat==0.6.3
pycparser==2.19
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
ruamel-yaml==0.15.87
s3transfer==0.3.3
six==1.14.0
tqdm==4.36.1
urllib3==1.25.8
Werkzeug==1.0.0
